In progress

+++ metadata-io/src/main/java/com/linkedin/metadata/entity/ebean/EbeanRetentionService.java
@@ -19,10 +19,10 @@ import static com.linkedin.metadata.Constants.ASPECT_LATEST_VERSION;
 
 
 @RequiredArgsConstructor
-public class EbeanRetentionService implements RetentionService {
+public class EbeanRetentionService extends RetentionService {
   private final EntityService _entityService;
   private final EbeanServer _server;
-  private final Clock _clock = Clock.systemUTC();;
+  private final Clock _clock = Clock.systemUTC();
 
   public EntityService getEntityService() {
     return _entityService;
@@ -41,20 +41,23 @@ public class EbeanRetentionService implements RetentionService {
         .eq(EbeanAspectV2.URN_COLUMN, urn.toString())
         .eq(EbeanAspectV2.ASPECT_COLUMN, aspectName)
         .ne(EbeanAspectV2.VERSION_COLUMN, ASPECT_LATEST_VERSION);
+    boolean retentionApplied = false;
     for (Retention retention : retentionPolicies) {
-      if (retention.hasVersion()) {
-        deleteQuery = applyVersionBasedRetention(urn, aspectName, deleteQuery, retention.getVersion(), retentionContext.flatMap(
-            RetentionService.RetentionContext::getMaxVersion));
-      } else if (retention.hasTime()) {
-        deleteQuery = applyTimeBasedRetention(deleteQuery, retention.getTime());
+      if (retention.hasVersion() && applyVersionBasedRetention(urn, aspectName, deleteQuery, retention.getVersion(),
+          retentionContext.flatMap(RetentionService.RetentionContext::getMaxVersion))) {
+        retentionApplied = true;
+      } else if (retention.hasTime() && applyTimeBasedRetention(deleteQuery, retention.getTime())) {
+        retentionApplied = true;
       }
     }
 
-    deleteQuery.delete();
+    // Only run delete if at least one of the retention policies are applicable
+    if (retentionApplied) {
+      deleteQuery.delete();
+    }
   }
 
-  private long getMaxVersion(@Nonnull final String urn,
-      @Nonnull final String aspectName) {
+  private long getMaxVersion(@Nonnull final String urn, @Nonnull final String aspectName) {
     List<EbeanAspectV2> result = _server.find(EbeanAspectV2.class)
         .where()
         .eq("urn", urn)
@@ -68,20 +71,22 @@ public class EbeanRetentionService implements RetentionService {
     return result.get(0).getKey().getVersion();
   }
 
-  protected ExpressionList<EbeanAspectV2> applyVersionBasedRetention(@Nonnull Urn urn, @Nonnull String aspectName,
+  protected boolean applyVersionBasedRetention(@Nonnull Urn urn, @Nonnull String aspectName,
       @Nonnull final ExpressionList<EbeanAspectV2> querySoFar, @Nonnull final VersionBasedRetention retention,
       final Optional<Long> maxVersionFromUpdate) {
     long largestVersion = maxVersionFromUpdate.orElse(getMaxVersion(urn.toString(), aspectName));
 
-    if (largestVersion == 0) {
-      return querySoFar;
+    if (largestVersion < retention.getMaxVersions()) {
+      return false;
     }
-    return querySoFar.le(EbeanAspectV2.VERSION_COLUMN, largestVersion - retention.getMaxVersions() + 1);
+    querySoFar.le(EbeanAspectV2.VERSION_COLUMN, largestVersion - retention.getMaxVersions() + 1);
+    return true;
   }
 
-  protected ExpressionList<EbeanAspectV2> applyTimeBasedRetention(
-      @Nonnull final ExpressionList<EbeanAspectV2> querySoFar, @Nonnull final TimeBasedRetention retention) {
-    return querySoFar.lt(EbeanAspectV2.CREATED_ON_COLUMN,
+  protected boolean applyTimeBasedRetention(@Nonnull final ExpressionList<EbeanAspectV2> querySoFar,
+      @Nonnull final TimeBasedRetention retention) {
+    querySoFar.lt(EbeanAspectV2.CREATED_ON_COLUMN,
         new Timestamp(_clock.millis() - retention.getMaxAgeInSeconds() * 1000));
+    return true;
   }
 }
